Add a cast when converting from uint64_t to int
This silences warnings with MSVC.

+++ codec/decoder/core/src/decoder.cpp
@@ -993,8 +993,8 @@ void UpdateDecStatNoFreezingInfo (PWelsDecoderContext pCtx) {
     ResetDecStatNums (pDecStat);
     pDecStat->iAvgLumaQp = iTotalQp;
   } else
-    pDecStat->iAvgLumaQp = (uint64_t) (pDecStat->iAvgLumaQp * pDecStat->uiDecodedFrameCount + iTotalQp) /
-                           (pDecStat->uiDecodedFrameCount + 1);
+    pDecStat->iAvgLumaQp = (int) ((uint64_t) (pDecStat->iAvgLumaQp * pDecStat->uiDecodedFrameCount + iTotalQp) /
+                           (pDecStat->uiDecodedFrameCount + 1));
 
   //update IDR number
   if (pCurDq->sLayerInfo.sNalHeaderExt.bIdrFlag) {
